+++ services/workers/data-fetcher-2.0/src/DataFetcher.Worker/Application/Providers/Alpaca/AlpacaStockBackfillService.cs
@@ -78,7 +78,7 @@ public class AlpacaStockBackfillService : IAlpacaStockBackfillService
                     HighPrice = (decimal)bar.High,
                     LowPrice = (decimal)bar.Low,
                     ClosePrice = (decimal)bar.Close,
-                    Volume = bar.Volume
+                    Volume = (long)bar.Volume
                 }).ToList();
 
                 await _priceRepository.UpsertStockPricesBatchAsync(prices);
