+++ services/workers/data-fetcher-2.0/src/DataFetcher.Worker/Domain/Providers/Alpaca/Models/AlpacaBar.cs
@@ -20,7 +20,7 @@ public class AlpacaBar
     public double Close { get; set; }
 
     [JsonPropertyName("v")]
-    public long Volume { get; set; }
+    public double Volume { get; set; }
 
     [JsonPropertyName("n")]
     public long TradeCount { get; set; }
